Upgrade to Micronaut AOT 1.0.0-M6
In Micronaut AOT 1.0.0-M6, loading the static optimizations has
been reworked and needs to be done using a dedicated API. This is
what this commit fixes. The API change was needed to make sure
that the optimizations are set _before_ they are actually read.
There was a possibility that the optimizations are set after,
typically in AWS functions.