Make sensitive endpoints error by default
We could go further, and allow users to add some configuration to check custom roles,
but as users have to create a replacement for `SensitiveEndpointRule` themselves, this
is something that can more simply be done in that custom replacement.

It also would make the API change much more breaking.

As this is a change in functionality, I have tagged this as `3.4.0`