Fix tests
Since https://github.com/micronaut-projects/micronaut-core/pull/7367 any Cors request returns a 404
if the method and path does not exist.  This was the case here.

As the Cors filter doesn't handle the request when it's a 404, the request bubbles down to the security
filter, which fails it as there's no authorization (correctly for Cors requests I believe).

The fix is to add a path with the expected method