Avoid initializing logging at build time for GraalVM (#1171)
Custom conditions are encoding into the AnnotationMetadata which is initialized at build time, for this reason conditions can never use logging safely otherwise they will initialize the log infrastructure. This PR moves the logging in AuthenticationModeCondition to a static inner class which avoids initializing the logging when the class is initialized.