doc: idtoken not custom OpenIdAuthenticationMapper (#1408)
This PR documents that it is not possible to use `micronaut.security.authentication` = `idtoken` with a custom `OpenIdAuthenticationMapper`. When you use a custom `OpenIdAuthenticationMapper` is because you wish to modify the originall idtoken and add for example extra claims. That requries generate a new token. When you use `idtoken`, you say that you want to use the OAutheServer `idtoken` directly

Close #1391