Fix Keycloak docker tests for 4.6.x
We used to use an elderly version of jboss/keycloak for running our Keycloak tests.

The jboss/keycloak docker image is no longer available, so we need to do something to re-instate the tests.

Instead of using the x86 keycloak/keycloak image, this PR switches to the latest (at time of writing) bitnami/keycloak image.
These images have arm versions, so we can get rid of a lot of the complications.

Updating to the latest has raised 2 issues with our current Keycloak support.

1. Keycloak 17+ got rid of the /auth prefix to their URLs.  In this PR I added it back via an environment variable, but we should support some configuration to support this.
2. The logout flow for Keycloak has changed, so we have to enable legacy flow. Basically the redirect_uri parameter has been deprecated, so we need to capture and use some keys to go back to the previous behavior.

I will raise issues for these 2 problems.

This commit also renames a package that was named incorrectly `keycloack`