Offload JWK Set fetching from the event loop
JwtTokenValidator is modified to explicitly execute
JwtValidator.validate in a separate thread from the Netty event loop.

JwtValidator.validate is blocking in its current implementation and
this allows it to be executed in a non-blocking fashion.

This is a short-term fix for #1633, where the preferred long-term fix
would be to refactor the JwtValidator API to be non-blocking.